script: Normalise how temporary buffers are freed
The temporary vr_buffer to hold the list of commands was being copied
into the vr_script and then freed whereas the extensions buffer was
being taken over by the vr_script on success or freed on failure. This
normalises them both to just always let vr_script take over the
buffers. This will free them anyway on error when the vr_script is
freed.