Add support for setting simple properties on the pipeline
Most of the properties that are just a simple integer, float, boolean
or enum value in the VkGraphicsPipelineCreateInfo struct or one of its
member structs can now be set directly with a line in the [test]
section. All of the properties set form part of the vr_pipeline_key so
they will cause a different pipeline to be created if multiple draw
calls are set with different values.

The possible properties are described in a header file using some
macro magic. This is used to generate the struct containing the values
as well as some tables containing the various offsets to store the
value in the create info structs.