Load the framebuffer contents after subsequent renders
vr_window now creates two render passes. One has the loadOp set to
DONT_CARE and the other sets it to LOAD. If a mix of probes and draw
calls are used in the test then it will now use the second one to
retain the framebuffer contents when starting the render again.