script: Add a seperate section for initialisation commands
The commands in the [test init] section are added to the list of
commands as normal but there is a seperate counter to keep track of
how many of them there are. The tester uses this to run just these
commands once during vr_test_start before the first frame. This can be
used to initialise buffers.