error-message: Add a callback to replace the error reporting
vr_config can now contain a pointer to a callback function to invoke
whenever an error is to be reported. If set fv_error_message will
invoke it instead of printing to stderr. The intention is that this
will make it easier to integrate VkRunner into other testing suites.

This requires touching quite a lot of files in order to pass the
vr_config along.