examples: added mat2x2 test, that tweaks matrix stride
It seems that by default, glslang validator always sets
matrix_stride=16, or at least it uses it for mat2x2.

But for a mat2x2 you could use matrix_stride=8. For that reason this
test includes directly the spirv assembly, as it needs to tweak the
stride.

Additionally, VkRunner doesn't use the explicit matrix_stride to
upload the data or to set the buffer size. We workaround that issue by
uploading a vec4.