subprocess: Buffer the output and pass it to vr_error_message
Instead of letting the subprocess directly write to the same stdout
and stderr as VkRunner, they are now redirected into a buffer.
Whenever a complete line is encountered it will be sent to
vr_error_message. That way an application using the VkRunner library
can also intercept the compiler output.