Add a vr_source object to represent a script source
Instead of having a list of files in the vr_config, the shader source
is now specified in a vr_source object. This also contains the token
replacements. The intention is to move away from having to reconfigure
the config whenever a new set of sources is used with the executor.