Remove vr_config from the public API
vr_config is now just an internal struct which is owned by the
vr_executor. All of the public API to set properties on the config are
now just directly set on the executor. This avoids the problem that
the lifetime of a vr_window was tied to the lifetime of the executor
but it was maintaining a pointer to the vr_config and the application
was free to use multiple configs to use different sets of source
files.