Move the descriptor pool to the vr_pipeline
Since 314f6467f the descriptor pool was freed by the vr_context but
created by the vr_pipeline. That meant that if the context was used
with multiple pipelines it would create multiple descriptor pools and
the earlier ones would be leaked. This was causing an error from the
validation layers.