Add a unit testing framework
Unit tests can now be defined in any source file using the
VR_UNIT_TEST macro. The macro is grepped for during the build process
to search for available unit tests. These are added to a global list
using the gcc constructor attribute. The tests are not built by
default but instead are conditionally built if BUILD_VKRUNNER_TESTS is
defined.

Fixes: #25