gui-worker: Encapsulate script execution in a separate process
Instead of directly invoking vr_executor_execute, it now forks a
separate process to do it and communicates with it over a pair of
pipes. This fixes a problem that if the script causes the process to
crash it would otherwise also bring down the UI. The process is
recycled so that it can still use the same vr_executor multiple times
if it doesn’t crash.

Because it is now using the Unix API, it no longer tries to build the
GUI on non-Unix OSs.