Use strtod_l to force the C locale
The normal strtod is locale-sensitive which means that on locales that
don’t use the full stop (.) as a separator it won’t be able to parse
doubles in the script properly. This patch makes it create a C locale
on the vr_config and use it via strtod_l. This function is a GLIBC
extension so it is conditionally compiled when using GLIBC.

Fixes #34