Make a struct to contain the two formats used in vr_window
Previously both the script and the window contained seperate pointers
to the color and depth/stencil formats for the window. These two have
now been encapsulated into a struct called vr_window_format. This will
be useful to add more window format properties such as the size in
later commits.