android: Replace android_test example with just building vkrunner exe
Removes the android_test example. This was built as a library with a
hardcoded example shader and it needed to be linked to something else
in order to run. Instead, we can just build the normal vkrunner
executable which can be copied onto a device to run any script over
the command line. This is probably more useful than android_test.

Reviewed-by: Samuel Iglesias Gonsálvez <siglesias@igalia.com>