Add a script to test the build
This just makes a build, installs it to a local prefix and then tries
to build the example from the README with it. This can be run to make
sure the example is still valid and to check that the public headers
are all correctly installed.

If we ever add unit tests it would be good to add them here too and
eventually integrate it into a CI system.