android: Move config.h to a static file
Previously the config.h was ‘generated’ from config.h.android in
Android.mk. However this generation was actually just calling ‘cp’ to
copy the file. It also had the awkward problem that the generated file
was placed in the source directory rather than the build directory.
Seeing as the file doesn’t really need to be generated, let’s just put
it in a directory in the source tree where the Android build can find
it and add that to the include path.