android: Add a dummy vr-vulkan-header.h
When building with CMake this file is auto-generated with an option to
replace the name of the Vulkan header. This file wasn’t being
generated in the Android build so the build would fail. This patch
just adds a static header with the default header name into the
android-only include directory.