script: Fix undefined behaviour treating struct members as array
Previously the parse_floats functions and friends were being used in a
few places to parse all the members of a struct as if the struct were
an array of floats. This relies on undefined behaviour assuming that
the compiler puts the same amount of padding between struct members as
it does for array members. To be on the safe side and to avoid
potential static analysis errors, this commit replaces them all with a
temporary array that gets copied to the struct values.

Fixes #18