test: Reset bindings after starting command buf, not render pass
Previously the current pipeline and buffers were marked as unbound
when starting a new render pass. This has now been moved further up to
be when the command buffer is started. This fixes a problem with
compute shaders if a probe command is issued between two dispatch
commands. The probe command makes it end the command buffer and the
second dispatch command starts it again. At that point VkRunner
believed the pipeline was still bound so it wouldn’t rebind it and the
test would fail.