box: Add a layout argument to the type query functions
The functions that query the size and alignment of box types now take
an extra argument to specify the layout of the type. This is stored in
a struct. Currently the only member represents whether the layout is
std140 or std430. Eventually it could also contain information on the
matrix major axis and strides.

vr_script now has members in the load state for the current layouts.
These are currently always std140 so there are no functional changes.