box: Fix muddling of base_alignment and strides
Previously the base alignment value was being incorrectly calculated
as both the matrix stride and array stride. Instead it is supposed to
just be the alignment of one column of the type and is not rounded up
to the size of a vec4.

The rounding up to a vec4 has now been moved into the matrix stride
calculation. A new function is added to get the array stride by
multiplying the matrix stride by the number of columns.

I don’t think this makes any practical difference but the wording of
the function names is much clearer.