script: Make the test line parsing functions more uniform
Previously some of the functions to parse a test line returned ‘false’
to mean the line does not match the command and the parser should
continue looking. Others checked before calling the function whether
the line belongs to this command and a ‘false’ return value meant that
there was an error parsing it.

This patch makes them all return an enum with the three possible
meanings instead. They all now check whether the line looks
appropriate for the command within the function.

In addition, some of the test lines do not add a command. Previously
these were divided into two sections in process_test_line so that any
lines that added a command had to be in the second half. This made
parsing the two “ssbo” commands a bit more difficult because they have
the same beginning signature but one adds a command and the other
doesn’t. Instead, there is now a helper function to add a command and
each processing function can decide whether to add a command or not
itself.