Add support for descriptor set of arrays
So now you can do things like this:

   ubo 1:5:3 subdata float 0 12

That will fill up the ubo with set 1, binding 5, array index 3. More
details on the updated README.md

It is worth to note that this implementation assigns one VkBuffer per
ubo component, so we added some extra descriptor data on vr_script to
handle which buffers are included on each descriptor.

Technically, you could have used just one buffer per ubo/ssbo. But in
order to do that the VkDescriptorBufferInfo offset for each subsequent
array components needs to be aligned with
minUniformBufferOffsetAlignment or minStorageBufferOffsetAlignment,
that would depend on the driver, so it would have been complicated to
set the offsets on the test.

FIXME: pending to update process_probe_ssbo_command now that we have
       array_index