test: Flush the entire buffer when modifying a UBO or SSBO
Instead of trying to flush a small memory range after updating a UBO
or an SSBO, it now just marks the buffer as dirty and then flushes the
entire buffer just in time before the command buffer is submitted.

This fixes a problem that the offset and size in the range is meant to
be a multiple of VkPhysicalDeviceLimits::nonCoherentAtomSize. It would
be possible to fix the flushed range to be a multiple of this, but in
the likely event that the script contains multiple small buffer
updates, it seems like it would be nicer to just flush the entire
buffer once.

Fixes #54