test: Flush the command buffer before setting buffer subdata
Previously if a script contains something like this:

uniform ubo 3 float 0  1.0
compute 1 1 1
uniform ubo 3 float 0  2.0
compute 1 1 1

then both computes would see the value 2.0 in the UBO. This is because
the memory in the buffer is just updated by writing into the
memory-mapped view of it and the commands in the command buffer are
only submitted at the end of the script or before a probe command.
That means that the command buffer would only be submitted after
updating the UBO a second time.

This patch changes it to also flush the command buffer when updating a
buffer. This has the result that you might expect where each compute
command is executed in a separate submission and each will see
different values in the command buffer.

See #46