Add a utility for tracking extension and feature requirements
Adds vr_requirements which keeps a list of required extension names
and a list of structures of features. A requirement can be added to
the struct just by adding its name. If the name matches a feature then
the appropriate struct will be added to the list, otherwise it will be
assumed to be the name of an extension.