Update assert for number of descriptor sets
We has an assert that checked that pipeline->n_desc_sets and
script->n_descriptors had the same value (and a comment suggesting
that were duplicated code).

This commit updates the assert, as that value can be different. In the
way descriptor sets are declared through the Vulkan API, you can't
have "gaps" when declaring it. So if your shader uses a block with
set=1, you still need to declare set = 0 (it is not like the binding).

So:

  * pipeline->n_desc_sets takes that into account, and represent the
    descriptor sets that are needed to be declared (allocated).

  * script->n_descriptors are the number of descriptors used when
    filling the data on the script (and assumes that the same used on
    the shader).

This commit adds also explanation of all this.