Add some utilities to parse integers
The utilities are like the ones in the hex crate but to parse integers
instead of floats. It doesn’t look like there’s an equivalent of strtol
in rust that can automatically deduce the radix from the prefix and also
return a tail pointer. These utilities try to implement that.