pipeline-key: Make return value from get_entrypoint caller owned
This will make it easier to return a string from the Rust port that has
a null terminator. Once the conversion of vr-pipeline to rust is
finished then it will be able to use a rust method call directly which
will just return a string slice without the extra allocation.