executor: Convert to Rust
This is more of a temporary conversion in order to help port Window. Not
all of the dependencies of the executor have been ported so it still has
to use some of the C API from Rust. The main reason to do this is so
that the Window code will be able to take an Rc<Context> and keep a
reference to it.