requirements: Remove the check for vkEnumerateInstanceVersion
I think this probably doesn’t make sense because at the point we are
checking we have already created the device and we can just check the
device version. I think the instance level version would only matter if
we wanted to use instance-level functions from a specific version, but
VkRunner doesn’t do that and nothing in the scripts can make it do so.
It also avoids the problem that the library isn’t available if the
context was created externally.