Build both a static lib for C and a rust lib for rust apps
In order to make the Rust API accessible from a program written in Rust,
the build scripts now build the vkrunner library a second time as a Rust
library and install that as well. Meson doesn’t seem to work if I try
access the Rust API from the static lib and we need the static lib to be
able to build C programs.