vk: Add a function to implement VK_NULL_HANDLE
This is used instead of using ptr::null_mut() everywhere where
VK_NULL_HANDLE would be used in C. VK_NULL_HANDLE is a #define so it
doesn’t appear in the bindgen output. On 32-bit platforms, the
“non-dispatchable” handles are defined as 64-bit integers instead of
pointers. This was breaking the build because the pointer returned from
null_mut can’t be implicitly converted to a u64. Instead, there is now a
new vk::null_handle function which maps to ptr::null_mut() if the
pointer size is 64-bit, or 0u64 otherwise.